move util functions into its own file
